Quotation Type Bug
Fixed a bug due to the type of quotations used. The code used in the generate_agent function of lib/listeners/http.py incorrectly matched the $Profile variable in data/agent/agent.ps1. This cause the generated agent not to be updated with the Empire listener's DefaultProfile values before being sent to the client. Changed the quotations in agent.ps1 to match the quotation in the generate_agent code = code.replace('$Profile..... code