Fix #749 - Add check for 'Agent' in self.module.options
When attempting to switch Modules it will try to set the current Agent for the
new module that will be used. However, this fails when switching from
`external/generate_agent.py` because it does not have this option within the
self.module.options.

Therefore, I changed it to check if the Key exists within self.module.options
and if it does not exist it will be set to ''.