Fixed exceptions and suppressed extraneous output in agents.py and http.py
When running background modules, when the agent reported in, a database lookup was performed for "task=0" for that agent, which did not exist in the database. This was causing an exception that was written to the terminal, so that every 5 seconds (or, beacon interval) the terminal I/O was being interleaved with an error message. The database lookup was part of unused code, so that block has been removed.

Additionally, the agents checking in also report the following:
```
[*] Agent SIOB3GNU returned results.
[*] Valid results returned by 172.17.0.1
```

Those messages do not provide useful information and again, cause interleaved I/O after each beacon. The messages have been suppressed by setting "print" to False for the dispatch events